requirements.txt
# LangExtract Test Requirements
# Generated from successful installation on macOS ARM64
# Python 3.11+ required

# Core LangExtract dependencies
absl-py==2.3.1
aiohttp==3.12.15
google-genai==1.37.0
pydantic==2.11.9
pydantic-core==2.33.2
python-dotenv==1.1.1
requests==2.32.5

# Data processing dependencies
numpy==2.3.3
pandas==2.3.2

# Development and code quality tools  
isort==6.0.1
pyink==24.3.0
pylint==3.3.8

# Additional dependencies for full functionality
async-timeout==5.0.1
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
attrs==25.3.0
exceptiongroup==1.3.0
frozenlist==1.7.0
multidict==6.6.4
propcache==0.3.2
yarl==1.20.1
typing-extensions==4.15.0
more-itertools==10.8.0
ml-collections==1.1.0
tqdm==4.67.1
PyYAML==6.0.2

# Authentication and networking
google-auth==2.40.3
httpx==0.28.1
httpcore==1.0.9
certifi==2025.8.3
urllib3==2.5.0
idna==3.10
anyio==4.10.0
sniffio==1.3.1
h11==0.16.0

# For date/time handling
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2

# For visualization features
jinja2==3.1.6
MarkupSafe==3.0.2

# Note: LangExtract itself should be installed via:
# git clone https://github.com/google/langextract.git
# cd langextract
# pip install -e ".[dev]"